requirements.txt
# torch 
# pip install torch==2.2.2 torchvision==0.17.2 torchaudio==2.2.2 --index-url https://download.pytorch.org/whl/cu121
xformers==0.0.25.post1
omegaconf==2.3.0
accelerate==0.28.0
einops==0.7.0
opencv_python==4.9.0.80
scipy==1.15.2
ftfy==6.2.0
regex==2023.12.25
python-dateutil==2.9.0.post0
timm==0.9.16
pytorch-lightning==2.2.1 # only for loading pretrained sd weight
tensorboard
protobuf==4.25.3 # for tensorboard
lpips==0.1.4
facexlib==0.3.0
gradio==4.43.0
polars==1.23.0
torchsde==0.2.6
bitsandbytes==0.44.1
# requirements for llava
transformers==4.37.2
tokenizers==0.15.1
sentencepiece==0.1.99
# requirements for ram
fairscale==0.4.4
pillow=9.1.0
numpy==1.26.4
